# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UIT Career Hub API
  version: 1.0.0
  description: Contract skeleton cho happy flow MVP của một trường UIT.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /companies:
    get:
      tags: [Students, Companies]
      operationId: listActivePartnerCompanies
      description: Danh bạ dành cho sinh viên; chỉ trả doanh nghiệp đối tác đang hoạt động và không chứa dữ liệu quản trị hoặc tài khoản recruiter.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: industry
          in: query
          schema: { type: string, maxLength: 120 }
        - name: hasRecruitingJobs
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Danh sách đối tác UIT đang hoạt động
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerDirectoryListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
  /companies/{companyId}:
    get:
      tags: [Students, Companies]
      operationId: getActivePartnerCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '200':
          description: Hồ sơ công khai của đối tác UIT đang hoạt động
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerDirectoryEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /companies/me/profile:
    get:
      tags: [Companies]
      operationId: getMyCompany
      responses:
        '200':
          description: Doanh nghiệp của tài khoản hiện tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
    patch:
      tags: [Companies]
      operationId: updateMyCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdateRequest'
      responses:
        '200':
          description: Hồ sơ đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
...
  /uit/companies:
    get:
      tags: [UIT]
      operationId: listPartnerCompanies
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Danh sách doanh nghiệp đối tác
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
    post:
      tags: [UIT]
      operationId: createPartnerCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateRequest'
      responses:
        '201':
          description: Doanh nghiệp đối tác đã tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyCreateResultEnvelope'
  /uit/companies/{companyId}:
    parameters:
      - $ref: '#/components/parameters/CompanyId'
    get:
      tags: [UIT]
      operationId: getPartnerCompany
      responses:
        '200':
          description: Hồ sơ đối tác và danh sách tài khoản tuyển dụng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
    patch:
      tags: [UIT]
      operationId: updatePartnerCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyAdminUpdateRequest'
      responses:
        '200':
          description: Hồ sơ đối tác đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
  /uit/companies/{companyId}/suspend:
    post:
      tags: [UIT]
      operationId: suspendPartnerCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyStateChangeRequest'
      responses:
        '200':
          description: Tạm ngưng đối tác, tài khoản liên quan và refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
  /uit/companies/{companyId}/reactivate:
    post:
      tags: [UIT]
      operationId: reactivatePartnerCompany
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyStateChangeRequest'
      responses:
        '200':
          description: Kích hoạt lại đối tác và các tài khoản bị ngưng theo doanh nghiệp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDetailEnvelope'
  /uit/companies/{companyId}/recruiters:
    post:
      tags: [UIT]
      operationId: createCompanyRecruiter
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRecruiterRequest'
      responses:
        '201':
          description: Tài khoản chờ kích hoạt; token chỉ trả/gửi một lần
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRecruiterCreateResultEnvelope'
  /uit/companies/{companyId}/recruiters/{userId}/suspend:
    post:
      tags: [UIT]
      operationId: suspendCompanyRecruiter
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Tạm ngưng tài khoản và thu hồi refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanyDetailEnvelope' }
  /uit/companies/{companyId}/recruiters/{userId}/reactivate:
    post:
      tags: [UIT]
      operationId: reactivateCompanyRecruiter
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200':
          description: Kích hoạt lại tài khoản do UIT tạm ngưng
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanyDetailEnvelope' }
  /uit/companies/{companyId}/recruiters/{userId}/activation-link:
    post:
      tags: [UIT]
      operationId: regenerateCompanyActivationLink
      parameters:
        - $ref: '#/components/parameters/CompanyId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Token thô chỉ trả một lần; database chỉ lưu SHA-256 hash
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompanyActivationResultEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CRON_SECRET
    refreshCookie:
      type: apiKey
      in: cookie
      name: uit_refresh_token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    CompanyId:
      name: companyId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Không có quyền thực hiện
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Không tìm thấy hoặc không được phép biết tài nguyên tồn tại
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Trùng lặp, idempotency conflict hoặc transition không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    UserStatus:
      type: string
      enum: [PENDING_ACTIVATION, ACTIVE, SUSPENDED, LOCKED]
...
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            traceId: { type: string }
...
    Company:
      type: object
      required: [id, code, name, partnerStatus]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        legalName: { type: [string, 'null'] }
        taxCode: { type: [string, 'null'] }
        industry: { type: [string, 'null'] }
        companySize: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'] }
        partnerStatus: { type: string, enum: [PENDING, ACTIVE, SUSPENDED, ENDED] }
        version: { type: integer, minimum: 1 }
        verifiedAt: { type: [string, 'null'], format: date-time }
        recruiterCount: { type: integer, minimum: 0 }
        activeRecruiterCount: { type: integer, minimum: 0 }
        recruitingJobCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PartnerDirectoryCompany:
      type: object
      required: [id, code, name, industry, companySize, description, website, address, verifiedAt, recruitingJobCount]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        industry: { type: [string, 'null'] }
        companySize: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'] }
        verifiedAt: { type: [string, 'null'], format: date-time }
        recruitingJobCount: { type: integer, minimum: 0 }
    PartnerDirectoryEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/PartnerDirectoryCompany' }
    PartnerDirectoryListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/PartnerDirectoryCompany' }
        meta: { $ref: '#/components/schemas/PageMeta' }
    CompanyRecruiter:
      type: object
      required: [userId, email, fullName, isPrimary, status, createdAt]
      properties:
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        title: { type: [string, 'null'] }
        isPrimary: { type: boolean }
        status: { $ref: '#/components/schemas/UserStatus' }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    CompanyDetail:
      allOf:
        - $ref: '#/components/schemas/Company'
        - type: object
          required: [recruiters]
          properties:
            recruiters:
              type: array
              items: { $ref: '#/components/schemas/CompanyRecruiter' }
    CompanyCreateRequest:
      type: object
      required: [code, name, primaryRecruiter]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]+$' }
        name: { type: string, maxLength: 180 }
        legalName: { type: [string, 'null'], maxLength: 255 }
        taxCode: { type: [string, 'null'], maxLength: 50 }
        industry: { type: [string, 'null'], maxLength: 120 }
        companySize: { type: [string, 'null'], maxLength: 80 }
        description: { type: [string, 'null'], maxLength: 5000 }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'], maxLength: 500 }
        primaryRecruiter: { $ref: '#/components/schemas/CompanyRecruiterRequest' }
    CompanyUpdateRequest:
      type: object
      required: [name, expectedVersion]
      properties:
        name: { type: string, maxLength: 180 }
        description: { type: [string, 'null'], maxLength: 5000 }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'], maxLength: 500 }
        industry: { type: [string, 'null'], maxLength: 120 }
        companySize: { type: [string, 'null'], maxLength: 80 }
        expectedVersion: { type: integer, minimum: 1 }
    CompanyAdminUpdateRequest:
      type: object
      required: [code, name, expectedVersion]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]+$' }
        name: { type: string, maxLength: 180 }
        legalName: { type: [string, 'null'], maxLength: 255 }
        taxCode: { type: [string, 'null'], maxLength: 50 }
        industry: { type: [string, 'null'], maxLength: 120 }
        companySize: { type: [string, 'null'], maxLength: 80 }
        description: { type: [string, 'null'], maxLength: 5000 }
        website: { type: [string, 'null'], format: uri }
        address: { type: [string, 'null'], maxLength: 500 }
        expectedVersion: { type: integer, minimum: 1 }
    CompanyRecruiterRequest:
      type: object
      required: [email, fullName]
      properties:
        email: { type: string, format: email }
        fullName: { type: string, maxLength: 180 }
        title: { type: [string, 'null'], maxLength: 120 }
    CompanyStateChangeRequest:
      type: object
      required: [expectedVersion, reason]
      properties:
        expectedVersion: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 5, maxLength: 1000 }
    ActivationToken:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string, minLength: 32 }
        expiresAt: { type: string, format: date-time }
...
    CompanyDetailEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/CompanyDetail' }
    CompanyCreateResultEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [company, activation]
          properties:
            company: { $ref: '#/components/schemas/CompanyDetail' }
            activation: { $ref: '#/components/schemas/ActivationToken' }
    CompanyRecruiterCreateResultEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [userId, company, activation]
          properties:
            userId: { type: string, format: uuid }
            company: { $ref: '#/components/schemas/CompanyDetail' }
            activation: { $ref: '#/components/schemas/ActivationToken' }
    CompanyActivationResultEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [company, activation]
          properties:
            company: { $ref: '#/components/schemas/CompanyDetail' }
            activation: { $ref: '#/components/schemas/ActivationToken' }
...
    ReasonRequest:
      type: object
      required: [reasonCode, note]
      properties:
        reasonCode: { type: string, minLength: 2, maxLength: 100 }
        note: { type: string, minLength: 5, maxLength: 2000 }
...